gcp/requirements.txt
tensorflow
google-cloud-storage
Pillow
flask
flask_cors